requirements.txt
pytest
sqlalchemy
flask
psycopg2-binary
requests
python-dateutil==2.8.2
